:ambulance: Change type of coordinates from integer to decimal

+++ database/migrations/20191117212713_create_tables.js
@@ -20,8 +20,8 @@ exports.up = function(knex) {
       table.increments("issue_id");
 
       table.text("description");
-      table.integer("longitude");
-      table.integer("latitude");
+      table.decimal("longitude");
+      table.decimal("latitude");
 
       table
         .integer("user_id")
